165727294-feature(index): update file structure
