+++ package.json
@@ -76,10 +76,7 @@
     "prettier": "^1.17.0",
     "sass-loader": "^7.1.0",
     "selenium-server": "^3.141.59",
-    "style-loader": "^0.23.1",
-    "webpack": "^4.30.0",
-    "webpack-cli": "^3.3.1",
-    "webpack-dev-server": "^3.3.1"
+    "style-loader": "^0.23.1"
   },
   "husky": {
     "hooks": {
