+++ src/assets/stylesheets/index.scss
@@ -1,3 +1,9 @@
+$font-stack: Roboto, Arial, Helvetica;
+$primary-color: #0E71EB;
+$secondary-color: #959595;
+
+
+
 .custom-center{
     text-align: center;
 }
