+++ src/components/presentations/Homepage/__test__/HomePage.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import { create } from 'react-test-renderer';
 import HomePage from '../Homepage';
-import HeroView from '../../../presentations/Hero-view/Heroview-presentations';
+import HeroView from '../../Hero-view/Heroview-presentations';
 
 describe('HomePage component', () => {
   it('should match snapshot', () => {
